DCMAW-9664: Move AsyncTokeStateConfiguration to async token directory
